feat: create satellite with cmc
