--- src/libs/shared/src/factory/mod.rs
@@ -1,3 +1,0 @@
-pub mod cmc;
-mod impls;
-mod types;
